Move latex table printing stuff to its own library
